# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NFL Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/teams/{team}": {
      "get": {
        "summary": "Get Team",
        "operationId": "get_team_teams__team__get",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2025,
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team}/roster": {
      "get": {
        "summary": "Get Team Roster",
        "operationId": "get_team_roster_teams__team__roster_get",
        "parameters": [
          {
            "name": "team",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RosterPlayer"
                  },
                  "title": "Response Get Team Roster Teams  Team  Roster Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/team-analytics": {
      "get": {
        "summary": "Get Team Analytics",
        "operationId": "get_team_analytics_team_analytics_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "title": "Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamAnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RosterPlayer": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "jersey_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jersey Number"
          },
          "headshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headshot Url"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "position",
          "jersey_number",
          "headshot_url"
        ],
        "title": "RosterPlayer",
        "description": "A player on a team's roster for a given season."
      },
...
      "TeamAnalyticsResponse": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "league": {
            "items": {
              "$ref": "#/components/schemas/TeamAnalyticsRow"
            },
            "type": "array",
            "title": "League"
          }
        },
        "type": "object",
        "required": [
          "season",
          "league"
        ],
        "title": "TeamAnalyticsResponse",
        "description": "/team-analytics?season=\u2026 returns the season plus per-team rows."
      },
...
      "TeamGame": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "week": {
            "type": "integer",
            "title": "Week"
          },
          "gameday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gameday"
          },
          "gametime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gametime"
          },
          "away_team": {
            "type": "string",
            "title": "Away Team"
          },
          "home_team": {
            "type": "string",
            "title": "Home Team"
          },
          "away_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Score"
          },
          "home_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Score"
          },
          "stadium": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stadium"
          },
          "roof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roof"
          },
          "surface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "temp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp"
          },
          "wind": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind"
          },
          "away_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Away Record"
          },
          "home_record": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Record"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "season",
          "week",
          "gameday",
          "gametime",
          "away_team",
          "home_team",
          "away_score",
          "home_score",
          "stadium",
          "roof",
          "surface",
          "temp",
          "wind",
          "away_record",
          "home_record"
        ],
        "title": "TeamGame",
        "description": "A team's schedule row inside /teams/{team}. Records walked forward\nby attach_records() are required (always set on this endpoint)."
      },
      "TeamLeader": {
        "properties": {
          "player_id": {
            "type": "string",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "jersey_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jersey Number"
          },
          "headshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headshot Url"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "attempts": {
            "type": "number",
            "title": "Attempts"
          },
          "completions": {
            "type": "number",
            "title": "Completions"
          },
          "pass_yards": {
            "type": "number",
            "title": "Pass Yards"
          },
          "pass_tds": {
            "type": "number",
            "title": "Pass Tds"
          },
          "interceptions_thrown": {
            "type": "number",
            "title": "Interceptions Thrown"
          },
          "sacks_taken": {
            "type": "number",
            "title": "Sacks Taken"
          },
          "pass_epa": {
            "type": "number",
            "title": "Pass Epa"
          },
          "carries": {
            "type": "number",
            "title": "Carries"
          },
          "rush_yards": {
            "type": "number",
            "title": "Rush Yards"
          },
          "rush_tds": {
            "type": "number",
            "title": "Rush Tds"
          },
          "rush_epa": {
            "type": "number",
            "title": "Rush Epa"
          },
          "targets": {
            "type": "number",
            "title": "Targets"
          },
          "receptions": {
            "type": "number",
            "title": "Receptions"
          },
          "rec_yards": {
            "type": "number",
            "title": "Rec Yards"
          },
          "rec_tds": {
            "type": "number",
            "title": "Rec Tds"
          },
          "air_yards": {
            "type": "number",
            "title": "Air Yards"
          },
          "yac": {
            "type": "number",
            "title": "Yac"
          },
          "rec_epa": {
            "type": "number",
            "title": "Rec Epa"
          },
          "solo_tackles": {
            "type": "number",
            "title": "Solo Tackles"
          },
          "assist_tackles": {
            "type": "number",
            "title": "Assist Tackles"
          },
          "tackles_for_loss": {
            "type": "number",
            "title": "Tackles For Loss"
          },
          "sacks": {
            "type": "number",
            "title": "Sacks"
          },
          "qb_hits": {
            "type": "number",
            "title": "Qb Hits"
          },
          "def_interceptions": {
            "type": "number",
            "title": "Def Interceptions"
          },
          "pass_breakups": {
            "type": "number",
            "title": "Pass Breakups"
          },
          "forced_fumbles": {
            "type": "number",
            "title": "Forced Fumbles"
          },
          "fumble_recoveries": {
            "type": "number",
            "title": "Fumble Recoveries"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "player_name",
          "position",
          "jersey_number",
          "headshot_url",
          "games_played",
          "attempts",
          "completions",
          "pass_yards",
          "pass_tds",
          "interceptions_thrown",
          "sacks_taken",
          "pass_epa",
          "carries",
          "rush_yards",
          "rush_tds",
          "rush_epa",
          "targets",
          "receptions",
          "rec_yards",
          "rec_tds",
          "air_yards",
          "yac",
          "rec_epa",
          "solo_tackles",
          "assist_tackles",
          "tackles_for_loss",
          "sacks",
          "qb_hits",
          "def_interceptions",
          "pass_breakups",
          "forced_fumbles",
          "fumble_recoveries"
        ],
        "title": "TeamLeader",
        "description": "Per-player season totals for one team. Returned twice on /teams/{team}:\nonce for regular season (leaders), once for playoffs (playoff_leaders)."
      },
      "TeamProfile": {
        "properties": {
          "team": {
            "type": "string",
            "title": "Team"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/TeamGame"
            },
            "type": "array",
            "title": "Games"
          },
          "leaders": {
            "items": {
              "$ref": "#/components/schemas/TeamLeader"
            },
            "type": "array",
            "title": "Leaders"
          },
          "playoff_leaders": {
            "items": {
              "$ref": "#/components/schemas/TeamLeader"
            },
            "type": "array",
            "title": "Playoff Leaders"
          }
        },
        "type": "object",
        "required": [
          "team",
          "season",
          "games",
          "leaders",
          "playoff_leaders"
        ],
        "title": "TeamProfile",
        "description": "Response for /teams/{team}?season=\u2026"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}